+++ src/shared/event/client_event.rs
@@ -300,7 +300,10 @@ impl ClientEvent {
                         "applying event `{}` from client `{client}`",
                         any::type_name::<E>()
                     );
-                    client_events.send(FromClient { client, event });
+                    client_events.send(FromClient {
+                        client_id: client.into(),
+                        event,
+                    });
                 }
                 Err(e) => debug!(
                     "ignoring event `{}` from client `{client}` that failed to deserialize: {e}",
@@ -335,7 +338,7 @@ impl ClientEvent {
                 any::type_name::<E>()
             );
             client_events.send_batch(events.drain().map(|event| FromClient {
-                client: SERVER,
+                client_id: ClientId::Server,
                 event,
             }));
         }
@@ -442,10 +445,10 @@ impl<E: Event> FromWorld for ClientEventReader<E> {
 /// Emitted only on server.
 #[derive(Clone, Copy, Event, Deref, DerefMut)]
 pub struct FromClient<T> {
-    /// Entity representing a connected client or [`SERVER`] that sent the event.
+    /// Sender of the event.
     ///
     /// See also [`ConnectedClient`].
-    pub client: Entity,
+    pub client_id: ClientId,
 
     /// Transmitted event.
     #[deref]